+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/MavenPublishBaseExtension.kt
@@ -364,8 +364,15 @@ abstract class MavenPublishBaseExtension(
     }
 
     when {
-      project.plugins.hasPlugin("org.jetbrains.kotlin.multiplatform") ->
-        configure(KotlinMultiplatform(defaultJavaDocOption(plainJavadocSupported = false)))
+      project.plugins.hasPlugin("org.jetbrains.kotlin.multiplatform") -> {
+        val variant = project.findOptionalProperty("ANDROID_VARIANT_TO_PUBLISH") ?: "release"
+        configure(KotlinMultiplatform(
+          javadocJar = defaultJavaDocOption(plainJavadocSupported = false),
+          sourcesJar = true,
+          androidVariantsToPublish = listOf(variant),
+          forceAndroidVariantsIfNotEmpty = false,
+        ))
+      }
       project.plugins.hasPlugin("com.android.library") -> {
         val variant = project.findOptionalProperty("ANDROID_VARIANT_TO_PUBLISH") ?: "release"
         configure(AndroidSingleVariantLibrary(variant))